Lynda/Datascience/Ex_Files_Deep_Learning_Face_Recog/Exercise Files/Mac Exercise Files/requirements.txt
dlib==19.8.1; sys_platform == 'windows'
dlib; sys_platform == 'darwin'
face_recognition==1.2.2
pillow